Moved connection in seperate file
